requirements.txt
scp==0.14.5
paramiko==3.3.1
python-dotenv==1.0.0